started adding user routes
