+++ server/db/index.js
@@ -15,8 +15,6 @@ User.hasMany(Cart);
 Man.hasMany(Tear);
 
 
-
-
 // if we had more models, we could associate them in this file
 // e.g. User.hasMany(Reports)
 module.exports =  db;
